test: fix broken test and types
